Clean comments
removed useless comments
